refactor: enlarge some modules
